fix(SniffRouter): harden lifecycle and split files
