+++ tunnels/SniffRouter/downstream/resume.c
@@ -1,0 +1,13 @@
+#include "structure.h"
+
+void sniffrouterTunnelDownStreamResume(tunnel_t *t, line_t *l)
+{
+    sniffrouter_lstate_t *ls = lineGetState(l, t);
+
+    if (ls->prev_finished || ls->next_finished)
+    {
+        return;
+    }
+
+    tunnelPrevDownStreamResume(t, l);
+}
