+++ tunnels/SniffRouter/instance/api.c
@@ -1,0 +1,8 @@
+#include "structure.h"
+
+api_result_t sniffrouterTunnelApi(tunnel_t *instance, sbuf_t *message)
+{
+    discard instance;
+    bufferpoolReuseBuffer(getWorkerBufferPool(getWID()), message);
+    return (api_result_t){.result_code = kApiResultOk};
+}
