+++ tunnels/SniffRouter/instance/destroy.c
@@ -1,0 +1,6 @@
+#include "structure.h"
+
+void sniffrouterTunnelDestroy(tunnel_t *t)
+{
+    tunnelDestroy(t);
+}
