+++ tunnels/SniffRouter/instance/index.c
@@ -1,0 +1,6 @@
+#include "structure.h"
+
+void sniffrouterTunnelOnIndex(tunnel_t *t, uint16_t index, uint16_t *mem_offset)
+{
+    tunnelDefaultOnIndex(t, index, mem_offset);
+}
