+++ tunnels/SniffRouter/instance/prepair.c
@@ -1,0 +1,6 @@
+#include "structure.h"
+
+void sniffrouterTunnelOnPrepair(tunnel_t *t)
+{
+    discard t;
+}
