+++ tunnels/SniffRouter/instance/start.c
@@ -1,0 +1,6 @@
+#include "structure.h"
+
+void sniffrouterTunnelOnStart(tunnel_t *t)
+{
+    discard t;
+}
